:memo: Use GitHub's URL shortener
Rawgit is not affiliated by GitHub. It's run by a single person in their
free time and has no guarantee for uptime or support [1].

Git.io on the other hand is a service provided and used by GitHub, so in
theory it should be safer to use for GitHub URLs. But it's also a much
shorter URL which is nicer for the user.

1: https://github.com/rgrove/rawgit/blob/master/FAQ.md#i-need-guaranteed-100-uptime-should-i-use-cdnrawgitcom
2: https://github.com/blog/985-git-io-github-url-shortener